+++ src/app/meal/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import MealPage from "@/templates/meal";
+import MealPage from "@/templates/meal/layouts";
 import React from "react";
 
 const Meal = () => {
