--- src/constants/meal.constant.ts
@@ -1,24 +1,0 @@
-const MEAL = {
-  PREV: {
-    NAME: "이전",
-    TYPE: "prev",
-  },
-  FIRST: {
-    NAME: "첫번째",
-    TYPE: "first",
-  },
-  SECOND: {
-    NAME: "두번째",
-    TYPE: "second",
-  },
-  THIRD: {
-    NAME: "세번째",
-    TYPE: "third",
-  },
-  NEXT: {
-    NAME: "다음",
-    TYPE: "next",
-  },
-};
-
-export default MEAL;
