+++ src/provider/layoutProvider.helper.tsx
@@ -10,6 +10,8 @@ import { ROUTER } from "@/constants";
 const LayoutProvider = ({ children }: React.PropsWithChildren) => {
   const { isWindow } = useWindow();
   const routerName = isWindow ? window.location.pathname : "";
+  const dontNeedAsidePage: Array<string> = [ROUTER.LOGIN, ROUTER.MEAL];
+  const isNeedAsidePage = !dontNeedAsidePage.includes(routerName);
 
   return (
     <>
@@ -20,7 +22,7 @@ const LayoutProvider = ({ children }: React.PropsWithChildren) => {
         <Header />
         <Main>
           <Container>{children}</Container>
-          {routerName !== ROUTER.LOGIN && <Aside />}
+          {isNeedAsidePage && <Aside />}
         </Main>
         <Footer />
       </Layout>
