--- src/templates/bamboo/interfaces/@props/index.ts
@@ -1,1 +1,0 @@
-export type { default as BambooPostListItemProps } from "./bambooListItem.interface";
