+++ src/templates/bamboo/interfaces/index.ts
@@ -1,1 +1,2 @@
 export type { default as BambooPostType } from "./bambooPost.interface";
+export type { default as BambooPostListItemProps } from "./@props/bambooListItem.interface";
