+++ src/templates/home/layouts/BasicMode/index.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { getMealName } from "@/helpers";
 import { Column, Row } from "@/components/Flex";
 import { Aside } from "@/components/common";
-import useMeal from "@/hooks/useMeal";
-import useUser from "@/hooks/useUser";
+// import useMeal from "@/hooks/useMeal";
+// import useUser from "@/hooks/useUser";
 import { useMainQuery } from "../../services/query.service";
 import HomeMeal from "./HomeMeal";
 import HomeReserve from "./HomeReserve";
